requirements.txt
flask
torch
torchvision

faiss-cpu
pillow
flask-cors
ftfy
regex
tqdm
